Added react-tostify package and added validation in laravel-react-api

+++ laravel-react-api/app/Http/Controllers/UserCtrl.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Validator;
 
 class UserCtrl extends Controller
 {
@@ -26,17 +27,41 @@ class UserCtrl extends Controller
      */
     public function store(Request $request)
     {
-        //
+        // $validator = Validator::make($request->all(), [
+        //     'name' => 'required|min:3|max:255',
+        //     'email' => 'required|email|max:255|unique:users',
+        //     'password' => 'required|min:6|max:20',
+        // ], [
+        //     'email.unique' => 'Email already exists. Try another.',
+        // ]);
+
+        $inputs = $request->all();
+        $rules = [
+            'name' => 'required|min:3|max:255',
+            'email' => 'required|email|max:255|unique:users',
+            'password' => 'required|min:6|max:20',
+        ];
+        $messages = [
+            'email.unique' => 'Email already exists. Try another.',
+        ];
+        $validator = Validator($inputs, $rules, $messages);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+
         $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
             'password' => bcrypt($request->password),
         ]);
 
-        return response()->json([
-            'message' => 'Successfully created user!',
-            'user' => $user
-        ], 201);
+        if ($user) {
+            // return response()->json($user, 201);
+            return response()->json(['message' => 'User created successfully.', 'user' => $user], 201);
+        } else {
+            return response()->json(['message' => 'User not created'], 400);
+        }
     }
 
     /**
@@ -47,6 +72,13 @@ class UserCtrl extends Controller
      */
     public function show($id)
     {
+        $validator = Validator::make(['id' => $id], [
+            'id' => 'required|exists:users,id',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
         //
         return User::find($id);
     }
@@ -61,6 +93,16 @@ class UserCtrl extends Controller
     public function update(Request $request, $id)
     {
         //
+        $validator = Validator::make($request->all(), [
+            'name' => 'min:3|max:255',
+            'email' => 'email|max:255',
+            'password' => 'min:6|max:20',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+
         $user = User::find($id);
         $user->update($request->all());
 
@@ -79,6 +121,14 @@ class UserCtrl extends Controller
     public function destroy($id)
     {
         //
+        $validator = Validator::make(['id' => $id], [
+            'id' => 'required|exists:users,id',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+
         $user = User::find($id);
         $user->delete();
